+++ Tweet.h
@@ -11,22 +11,21 @@ using namespace std;
 class Tweet
 {
     private:
-        int id; //Tweet ID
+        long id; //Tweet ID
         int sentiment; //For training data, the ACTUAL sentiment value.
         int predictedSentiment; //For assigning what the program thinks this tweet's sentiment is.
         
     public: 
         Tweet();
         void Initialize();
-        int getID();
+        long getID();
         int getSentiment();
         int getPredicted();
-        void setID(int);
+        void setID(long);
         void setSentiment(int);
         void setPredicted(int);
         vector<int> wordvec;
         vector<char> garbvec; // Holds all commas and stuff.
-
 };
 
 #endif
